+++ src/Bus/QueryBus.php
@@ -13,6 +13,7 @@ use SomeWork\CqrsBundle\Exception\MultipleHandlersException;
 use SomeWork\CqrsBundle\Exception\NoHandlerException;
 use SomeWork\CqrsBundle\Support\StampsDecider;
 use Symfony\Component\Messenger\Exception\HandlerFailedException;
+use Symfony\Component\Messenger\Exception\NoHandlerForMessageException;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Messenger\Stamp\StampInterface;
 
@@ -60,6 +61,8 @@ final class QueryBus implements QueryBusInterface
             $envelope = $this->bus->dispatch($query, $stamps);
         } catch (HandlerFailedException $exception) {
             throw SynchronousResult::unwrap($exception);
+        } catch (NoHandlerForMessageException $exception) {
+            throw new NoHandlerException($query::class, self::BUS_NAME, $exception);
         }
 
         $handledStamps = SynchronousResult::handledStamps($envelope, self::BUS_NAME);